--- src/device-uptime/models/raw_feeds_pms.py
@@ -1,4 +1,0 @@
-class RawFeedsPMS():
-
-    def __init__(self, tenant):
-        self.tenant = tenant
